+++ hooks/hook-pyarrow.py
@@ -1,0 +1,7 @@
+"""
+PyInstaller hook to exclude pyarrow (not needed in production)
+"""
+excludedimports = ['pyarrow']
+datas = []
+binaries = []
+hiddenimports = []