+++ hooks/hook-torchvision.py
@@ -1,0 +1,7 @@
+"""
+PyInstaller hook to exclude torchvision
+"""
+excludedimports = ['torchvision']
+datas = []
+binaries = []
+hiddenimports = []